requirements.txt
# Core data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Market data
yfinance>=0.2.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Interactive development
jupyter>=1.0.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0

pyarrow

torch

ib_insync

fredapi